Simplify logger a little big
